controller: send Ack in reply to Channel Preference Report
Currently, if the Channel Preference Report contains invalid
values, no Ack is sent to the Agent, causing the certification
test 5.16.1 to fail.
The spec mandates the Controller to send the Ack and doesn't mention
any check that could allow it not to send this frame.

Sending the Ack before parsing the Report allows to meet the standard
requirements without changing the parsing code.

Fixed PPM-2260

Signed-off-by: Sylvain LE ROUX <prplmesh.opensources@sagemcom.com>