controller: update TRIGGER_CHANNEL_SELECTION handling
Currently when we handle the TRIGGER_CHANNEL_SELECTION request in the
controller, we sometimes need to send a CHANNEL_PREFERENCE query to the
agent.
This causes a timeout when the BML internal tries to handle the NonVS
message CHANNEL_PREFERENCE_QUERY_MESSAGE.
It receives that message because the controller shares its cmdu_tx with
the tasks.

So we need to send the response back to the BML before we send the
events that trigger the NonVS messages.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>