tools: device: logs: add iptables and syslog
Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>

tools: device: logs: add iptables-filter.txt

In some cases (e.g. [1]) we can't SSH to the device to run tests.

After checking manually, it turns out the firewall prevents it, as the
INPUT chain has no rule and policy DROP.

To make it easier to debug such issues, add the content of the filter
table to the artifacts.

While at it, also add the syslogs.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2906575705

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>