tools: device: logs: add dm-Firewall.txt
For now we plainly remove the tr181-firewall service.

However, that causes a race condition where sometimes, the INPUT chain
of the filter table would be left with a DROP policy, and no rules.

SAH recommended that we instead keep the tr181-firewall service
running, but disabled.

To prepare for that, add the Firewall datamodel to the artifacts
already.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>