ci: configuration: enable tr181-firewall but accept anything on INPUT
Up to now there seem to be no way to disable the tr181-firewall (see
PCF-540).

The approach we've used up to now is to stop and disable the service
completely.
owever, that seem to cause a race condition at boot where sometimes,
the device would be left with a DROP policy on the INPUT chain and no
rules (e.g. in [1]). When this happens, tests cannot start because we
cannot access the device over SSH.

Until PCF-540 is solved, re-enable the tr181-firewall service, but set
the INPUT policy to ACCEPT.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2906575705

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>