bcl: add a wifiChannel class
In the current situation, we use a channel and channel-related
variables, like center frequency, bandwidth, and more,
as separate variables. It worked well because the 2.4ghz and 5ghz
bands have no overlapping channels. However, One of the problems of
the current situation of using separate variables is that the channel
number, the center frequency, and other variables may not be
correlated, and one may update one of them, but forget to
update the others. In addition, due to supporting a 6ghz band,
we can no longer depend on a channel to identify the band type
because the 5ghz and 6ghz bands have overlapping channels.

A wifiChannel class represents a channel of 2.4ghz, 5ghz, or 6ghz band,
which comprise of a channel number, center frequency, bandwidth, etc.

Benefits of using the class:
* Instead of declaring separate variables for a channel,
  center frequency, and bandwidth, one can declare this object only.
* Updating one attribute of the wifiChannel objects updates the other
  attributes respectfully. For example, changing a bandwidth of a
  wifiChannel object from 20mghz to 40mghz also calculates and updates
  the center frequency of the wifiChannel object.

Using this class allows you to describe a WifiChannel correctly,
and a more generic way to support future bands.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>