controller: send ACK in reply to Channel Scan Report
Currently, if the Channel Scan Report contains invalid values, no Ack is sent to the Agent,
resulting the failure of the certification test 5.14.2.

The spec mandates the Controller to send the Ack within 1 second after receiving the
Channel Scan Report and doesn't mention any check preventing the sending of this frame.

Sending the Ack before parsing the Report allows to meet the standard requirements without
changing the parsing code.

PPM-2263

Signed-off-by: Mahdi BAKLOUTI <prplmesh.opensources@sagemcom.com>