ci: add test-gw-repeater
Since we've been using docker-compose to run Boardfarm tests, we
dropped the usage of tests/test_gw_repeater.sh in CI.

It turns out that it can still be useful to run locally, to quickly
check if the onboarding works.

To make sure that it keeps on working, add a CI job for it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>