agent: capture timer period by-copy instead of by-reference
Because tasks_timer_period was declared within the scope of
thread_init(), when we leave the function the original memory is
released. This causes the lambda function to retrieve an invalid value
when we call to tasks_timer_period's count().
As can be seen from the issue described in PPM-2270.

To resolve this issue we can capture tasks_timer_period in the lambda
by-copy instead of by-reference, so even if the original memory is
released/corrupted. It will not affect the lambda's functionality.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>