agent: Add validation to prevent a crash
Where we iterate on the non-operable channels to print them,
a channel number and a comma are written to the output string variable,
and outside the loop, we pop out that last comma so the print would be
esthetic.
A crash could happen if the non-operable channels list is empty, and we
call anyway to "pop_back()".
The "std::string::pop_back()" documentation defines that calling that
method when the string is empty will end up with undefined behavior.

Add if statement to return 'None' string if the non-operable channels
list is empty.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>