ci: configuration: add a workaround for PCF-681
As described in PCF-681, sometimes the device doesn't have a
connectivity to the rest of the network (either because there is no
bridge, or the bridge or one of its interfaces is down, ...).

The problem seem to happen at the very first boot. With the PWHM
changes, we're going to have two different prplOS versions for every
pipeline (one that includes the PWHM and one that does not). As a
consequence, we're likely to see the issue much more often.

Until PCF-681 is fixed, try to work around the problem by restarting
tr181-bridging if needed.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>