common: Try to connect with the broker multiple times with a delay
When prplMesh processes are brought up, broker clients may try to
connect before the transport process is up.
On the client side, failure to connect the transport process means
FATAL.
To prevent random FATAL when prplMesh is started, return an error only
if the client failed to connect the transport for more than
'broker_client_connection_timeout' seconds.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>