ci: retry GL.iNet B1300 operational tests if needed
PCF-695 is currently blocking the merge of the PWHM changes ([1]) and
we don't have any runtime workaround for it.

One way to work around it would be to reboot the device and re-apply
the configuration. The problem with this approach is that we would
lose the logs of when the bug occurs, making it much more difficult to
debug.

Instead if the job fails, retry it entirely.  We have to do the same
for the non-PWHM operational tests on GL.iNet B1300, as the problem is
not dependent on which prplOS version is used but affects all current
prplOS versions at first boot (only one GL.iNet B1300 is available in
CI, so its prplOS version changes constantly).

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/3059

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>