agent: Fix crash on socket disconnect
CmduClient requires assigning callbacks when an event occurs.
On several prplmesh modules that have a client to another thread,
we have a call to "std::unique_ptr<CmduClient>::reset() to the pointer
that holds the client socket.
This causes to CmduClient destruction which leads to the destruction of
the callback function while we are inside it.
The result is undefined behavior which potentially leads to a crash.

Move the reset call of the client pointer outside of the callback
function.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>