controller: config: remove agent dependency
Removing
client_band_steering
client_optimal_path_roaming
client_optimal_path_roaming_prefer_signal_strength
client_11k_roaming
dfs_reentry
service_fairness_enabled

All options except the service_fairness are part of the
configuration exposed in the NbAPI under the
Device.WiFi.DataElements.Configuration path.

The removed lines are in an
if(local_master){} code block;

the local_master bool means the agent being onboarded
is the local one, i.e. running on the same device as
the controller;

this interaction between agent and controller config
disables certain features in the controller if the agent doesn't support
them, which makes sense in the case of an homogenous network
(i.e. all agents have the same configuration);

this interaction makes it so the configuration on the controller
expresses potential capabilities, and it's the local agent (or
option-specific bml messages) that activate or de-activate them;

this commit is a step towards the controller-only
mode of operation for prplmesh;

https://jira.prplfoundation.org/browse/PPM-2155

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>