controller: Fix unexpected Fatal
Fix unexpected Fatal which pushed as part of this MR: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/3150
The variable 'bssid_vap_map' is used to map BSSID to vap_id when and it
is filled when parsing the prplmesh vendor-specific TLV
"tlvVsBssidIfaceMapping".
A Fatal print was added in case there is a misalignment with AP
Operational BSS TLV.
The variable 'bssid_vap_map' is defined as a class member though it
should have been defined locally in the Topology Response handler.
Because it is class member, the data you pushed inside it remains after
the function ends.
So if there is a local agent on the Controller device, this map will
always be filled.
When a third-party Agent will send this message, the map will not be
empty, and BSSIDs reported AP Operational BSS TLV will not be there, so
it will reach a FATAL print.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>