ci: use ubus to disable DHCP clients and server
As reported by the SAH team, the managers service should not be
disabled.

Instead, use ubus to set their "Enable" to false.

In addition, we need to restore the DHCP services in case they were
removed by other pipelines (as we don't re-flash a new full image
unless the prplOS version on the device is different).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>