db/nbapi: clear method for DM remove all CAC Status reports
According to the specification, CACStatus is a group object,
so the method for clearing the datamodel should remove the
indexes of the CACStatus object (i.e. reports), not its subobjects.

At the moment we don't have a policy regarding the number
of reports stored in the DM - so, remove DM clearing.

PPM-1496

Signed-off-by: Victor Pushkarev <v.pushkarev@inango-systems.com>