ap-manager: refactor handling of WIFI_CREDENTIALS_UPDATE
As part of the changes required by PPM-2286, need to make the following
adjustments to the handling of WIFI_CREDENTIALS_UPDATE:
- Save the BSSID to the BWL struct
This will allow us to compare the BSSID to that of the hostapd config.
- Set teardown flag if needed
Using the teardown flag, we can skip the reconfiguration validation to
quickly teardown a VAP.
- Perform update_vap_credentials if bss_info_conf_list isn't empty.
If the incoming bss_info_conf_list is empty, we don't have a reason to
trigger the update_vap_credentials.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>