common: hostapd: don't set config_id if it's not supported
In case a "recent" prplMesh version is running on an older prplOS
version that does not support the config_id parameter, hostapd will
refuse to load the configuration file. For example:
"Line 57: unknown configuration item 'config_id'"

To keep backward compatibility with previous prplOS versions, only
update the config_id if it was already set before (by netifd or a
previous prplMesh configuration).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>