auto config task: add configuration validation function
As part of PPM-2286 we want to validate the the reconfiguration request
sent as part of the M2 messages.
To do this we would need a new function that will iterate over the
current BSSs and try to find its match in the incoming configuration.

If we find a similar config, we need to verify that a change is
actually needed.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>