auto config task: validate configuration before sending request
When we receive the AP_AUTOCONFIGURATION_WSC message in the auto config
task it should contain an M2 message. As part of PPM-2286 we want to
validate the the reconfiguration request sent as part of the M2
messages.

When we validate the config we remove any unnecessary reconfigurations.
If there are any reconfiguration remaining, we send the
WIFI_CREDENTIALS_UPDATE request to the AP-manager.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>