BWL: DWPALD: refactor the update_vap_credentials flow
Because of the changes required by PPM-2286, update_vap_credentials
no longer receives a list of all available vaps, instead it receives
a list of only the necessary changes from the current state as present
in the agent.

Thanks to this change we can refactor the update_vap_credentials flow
to follow the following logic:
- Iterate over the received configuration.
- Find the hostapd's BSS configuration.
- If the BSS is flagged for teardown, clear its configuration.
- Check if ssid, credentials & multiap configurations were changed.
- If the BSS is disabled in the hostapd, enable it.
Once we iterate over all the received configuration.
Need to check if the old configuration needs reconfiguration.
Save the changes to the hostapd file.
Send RECONF commands to any BSS that needs reconfiguration.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>