auto config task: expand validation lambdas
As part of PPM-2286 we want to validate the the reconfiguration request
sent as part of the M2 messages.

Expand the find_by_similarity & bss_needs_reconfiguration lambdas.
Still need to expand them farther, but we are limited by the
information stored in the AgentDB.

Once PPM-2296 is resolved, we can expand these lambdas.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>