agent: use controller preference instead of radio preference
In the select_next_channel function use controller preferences
instead of radio preferences for calculating the channel
preference in the channel_selection_task of the agent.
For the same purpose bring the controller preferences from
the structure sIncomingChannelSelectionRequest and made
changes to the function header of select_next_channel
function.

Closes https://jira.prplfoundation.org/browse/PPM-2297

Signed-off-by: badhri_s <bsrinivasan@maxlinear.com>