BWL: NL80211: refactor the update_vap_credentials flow
Because of the changes required by PPM-2286, update_vap_credentials
no longer receives a list of all available vaps, instead it receives
a list of only the necessary changes from the current state as present
in the agent.

Thanks to this change we can refactor the update_vap_credentials flow
to follow the following logic:
- Iterate over the received configuration.
- Find the hostapd's BSS configuration.
- Find the BSS in the available vaps.
- Update that specific configuration.
- Teardown any other configuration.
Save the changes to the hostapd file.
Send the UPDATE commands if needed.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>