prplmesh-hostapd: add update required to configuration
we can better understand if an update is required by checking this flag.

The update_required flag is initialized when we load the configuration
from the file, or store it to the file.
It is then changed when we set/update a value in the configuration.
Later we could compare it against the preexisting value for better
performance.

It cannot be directly modified, but using this added function we can
read it's state.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>