channel-selection: Refine primary channel handling
Remove some of the unnecessary logs from the channel-selection
flow as they are flooding the backhaul log file.
Also enhance some of the other logs to provide a clearer
image of the channel-selection process.
Refined the function find_best_beacon_channel which
would help the on demand channel selection process to
select higher bandwidths like 80 and 160 MHz.

Closes https://jira.prplfoundation.org/browse/PPM-2297

Signed-off-by: badhri_s <bsrinivasan@maxlinear.com>