agent: refactor cac_status_database class
Introduces the update_cac_status_db() method.

This method searches the Agent's DB for given radio and
fills in the parameter values for each channel type.

Handling all types of channels in one place is necessary
to avoid redundancy when filling CAC Status Report TLV
with different radios.

At the end, the method fills in map for a specific type
of channel - this is one of the ways to simplify the
exchange of values between methods and is necessary
for consistency with CacStatusInterface.

In the next commits, there will be processing of other
channel types.

Also, remove unused header.

PPM-1496

Signed-off-by: Victor Pushkarev <v.pushkarev@inango-systems.com>