tools: builder: rdk-b: split RDK-B and prplMesh builder images
Split the Dockerfile into two different stages so that it can be
re-used to build other RDK-B targets.

This allows developers to use the 'rdk-b-builder' target to manually
build for other RDK-B platforms (e.g. building the AirTies Easymesh
controller without prplMesh) on hosts where installing the required
packages locally is not possible.

For example, the first stage can be used to build an image like this:
docker build --build-arg UID="$UID" --build-arg GID="$(id -g)" --target rdk-b-builder --tag "rdk-b-builder-$USER" tools/docker/builder/rdk-b

The resulting image can then be used, for example with:
docker run -it -v "$(pwd)":"$(pwd)" --workdir "$(pwd)" "rdk-b-builder-$USER"

When the target stage is not specified the default is to use the last
one, so for other users of the Dockerfile (e.g. the CI), nothing
changes.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>