bwl: nl80211: remove max_num_sta
Currently when a new backhaul station wants to connect and there is
already a station connected, it fails with:
hostapd: no more room for new STAs (1/1)

When update_vap_credentials was first introduced in commit
4748b6dd3780606924b2ae5dd427fd967d081d35, it was copied and adapted
from the dwpal BWL.

In the dwpal BWL, new VAPs are created when new backhaul stations
associate to a backhaul AP. This might be the reason why max_num_sta
was set in the dwpal BWL: to avoid reaching the maximum number of
BSSes allowed by the radio.

With devices using the nl80211 BWL, no new VAP is created when a
backhaul station associates. There is thus no apparent reason to have
a hard limit on the number of stations, so remove it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>