Agent: Add BSSID TLV only on Profile3 (or above) Controller
WFA "tshark" version that is used on R2 testbed seems to not recognize
R3 TLVs, specifically the BSSID TLV, that was added on R3 to the Failed
Connection Message.
The tshark seems to treat the Failed Connection Message as
"malformed" so it doesn't pass the sniffer filter.

To prevent that, add the BSSID TLV only when the Controller is Profile3
or above.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>