BWL: dummy: refactor the update_vap_credentials flow
As part of PPM-2286 the auto configuration flow was refactored.
Because of this change the WIFI_CREDENTIALS_UPDATE request no longer
sends a full report of it's current BSSs status.
Instead the request now contains only BSS changes with correct a BSSID
parameter.

To comply with these changes we need to refactor the
update_vap_credentials function to follow this new information.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>