BWL: dummy: set the initial state of the vaps after attach
As part of PPM-2286 the auto configuration flow was refactored.
Because of this change the WIFI_CREDENTIALS_UPDATE request no longer
sends a full report of it's current BSSs status.
Instead the request now contains only BSS changes with correct a BSSID
parameter.

This means that the dummy BWL needs to behave like the actual hostapd
which starts with an initial VAP state.
To simulate this behavior, after we "attach" we want to set the initial
VAP status.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>