bwl: connect & roam: pair freq type with a channel
Pair a frequency type along with a channel in BWL's
connect and roam methods.
As part of supporting a 6GHz band, a channel alone is not
enough to identify the band, because 5GHz and 6GHz have
an overlapping channels.
Also, update the calls to those methods wherever they are called.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>