link_metrics_collection_task: Remove duplicated STA stats
When more than one BSS is configured on a radio, the stats of STA
connected on this radio are reported once for each BSS configured,
while they should be reported only once (with the BSS the STA is
connected to).

Filter STA stats according to the BSS they're connected when filling the
AP Metrics Response frame.

https://jira.prplfoundation.org/browse/PPM-2334

Signed-off-by: Sylvain Le Roux <prplmesh.opensources@sagemcom.com>