link_metrics_collection_task: order ap metrics queries by MID
If two AP Metrics Queries are received for the same BSSID, only one gets
a response. This can occur when an explicit query is sent while a
periodic one is already pending. This is for instance the case in
certification test 4.7.4

To avoid confusion between different queries, store the
query MID alongside the queried BSSID. This way, both queries can be
handled separately.

https://jira.prplfoundation.org/browse/PPM-2331

Signed-off-by: Sylvain Le Roux <prplmesh.opensources@sagemcom.com>