Add release.sh
This script creates a new release by creating a tag and updating the
changelog file accordingly.

Steps to create and push a release are now:

- Execute './release.sh <release_version>'.
- Check the result to make sure everything is as expected.
- Execute 'git push <release_version>'

Note that with the release creation workflow there was and still is a
small issue with the release time that is part of the changelog not
matching the actual release time (they are offset by a few seconds),
and the script does not try to fix it (i.e. the releases created by
the script have the same problem than if the release was created
manually).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>