tools: prplos: rebuild the full image
For the CI, we use a built prplOS tree to recompile prplMesh.

We cache the built prplOS tree using Docker, to be able to recompile
only what we need (for example, if the prplOS version didn't change,
we recompile only prplMesh).  This way, the compilation time for
prplOS devices is reduced to almost only compiling prplMesh.

After recompiling prplMesh we don't re-generate a full image, which
means the full image that we save (and make available in release
artifacts) contains the prplMesh version from the prplOS tree (usually
an older version).

This approach suits the CI well, as on test devices we anyway update
only prplMesh unless prplOS itself changed.

For users however, it can be confusing. Especially for releases for
example, users expect the full image available in the release
artifacts to contain the prplMesh version corresponding to the
release (and not the older one from the base prplOS tree).

To fix it, rebuild the full image instead of only prplMesh.  In order
to provide some information if the build fails, re-run it with V=sc,
one core and exit with failure if the parallel build failed.

Closes PPM-1620

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>