controller: vbss_task: remove misplaced error log
Currently, the VBSS task prints an error to the logs every time it
receives a 1905.1 message that it cannot handle.  However, during
normal execution the task receives any 1905.1 message received by the
controller, which leads to the logs containing many messages similar
to this one:
ERROR 09:37:55:177 <3069560160> vbss_task.cpp[82] --> task vbss_task id 8: Unknown CMDU message type: 0x6

These messages are confusing since they seem to indicate an error,
while it simply means the task received a message that it doesn't need
to handle.

To fix it do like for other tasks, return false if a message is not
handled and don't log it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>