controller: vbss_task: fix Client Capabilities TLV
When variable size members (e.g. SSID, password, DPP string, etc) are
set after addClass has already been called for another TLV, the length
of the first TLV ends up being incorrect.

This leads to a failing validation of the CMDU on the receiving
side (note the wrong size '1b 2f'):
ERROR 09:01:36:299 <beerocks_agent> broker_client_impl.cpp[221] --> Failed verifying CMDU!
DEBUG 09:01:36:299 <beerocks_agent> broker_client_impl.cpp[222] --> hex_dump (92 bytes):
00 00 80 40 00 2b 00 80 de 1b 2f 00 02 06 9a 94
73 43 c3 aa bb cc dd ee ff 0f 00 70 72 70 6c 4d
65 73 68 4e 65 74 77 6f 72 6b 0c 00 70 72 70 6c
6d 65 73 68 70 61 73 73 00 00 aa cc bb dd 00 00
00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
00 00 00 00 00 91 01 00 00 00 00 00

To fix it, call addClass for the client capabilities TLV only after
we're done filling in the Virtual BSS Creation TLV.

Note that there is a discrepancy between what the VBSS specs state and
what we do in practice.
The VBSS specs state:
If triggered to create a new VBSSs a radio on a Multi-AP Agent, a Multi-AP Controller shall send a Virtual BSS Request
message (see section 5.3) to the Multi-AP Agent, containing one Client Capabilities Report TLV and one Virtual BSS
Creation TLV (see section 6.2) with the Client Assoc field set to 0 for each new VBSS.
If triggered to move a VBSS from one radio on a Multi-AP Agent (the “source”) to another radio on the same or a different

In practice however, if the client is not associated yet, we don't
have a relevant association frame to include in the Client
capabilities TLV. We could include a frame from the previous
association, but it anyway couldn't be used by the agent to get the
client to move to the newly-created VBSS.
For this reason, only call addClass for the client capabilities TLV if
the client is associated.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>