agent: tlvf_utils: ignore 6GHz operating classes
This commit is necessary for the broadfarm-test to pass.

The function is based on a channel number to find
the operating class. With the introduction of 6Ghz band support,
we can no longer rely on a channel number because 6GHz channels
are overlapping with 2.4GHz, and 5GHz channels.
Thus, the support for a 6GHz band will be added in future commits.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>