bwl: nl80211_client: add a new method to add keys for a station
This new method will be used for the VBSS feature, to add the keys and
key sequences for a station when a VBSS is moved from one agent to
another one.

For now, no BWL implement it. In a follow-up commit, it will be
implemented for the shared nl80211_client_impl.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>