bwl: add new method to add keys for a station
In the future this method is going to be used by the VBSS feature to
add the new keys and key sequences for stations when they move from
one BSS to another one.

For now no BWL implemented it.

In a follow-up commit, and implementation based on nl80211_client will
be added.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>