agent: ap_manager: handle station and key creation when creating a VBSS
When creating a VBSS, check the client_assoc field to know whether we
have to add a new station for it or not.

If we do have to, use the newly-introduced ap_wlan_hal methods to
manually add the station from its association frame.

Once the station has been added, add pairwise and group keys for it.

Note that for now the encryption type is not known from the Virtual
BSS Creation TLV, so it's assumed to be CCMP. The actual encryption
type will be handled in a separate task.

Closes PPM-2292

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>