bwl: nl80211: move setting auth and encr parameters to a separate method
Setting hostapd parameters for authentication and encryption type can
be cumbersome, given the number of different cases that exists.

Since we will have to do the same when we manually add a single BSS,
prepare for it now by refactoring that part into a separate method. It
will be re-used in a follow-up commit to create single BSSes.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>