common: hostapd: add wrappers to set values for either head or vap
It can be cumbersome to have to have to explicitly use two different
methods to set parameters either for the head or for a vap, because in
many cases it doesn't matter whether the BSS to configure is the
'main' one or not.

Add two wrappers (for int and string values) to be able to use them
both for the head and vaps. If the vap is an empty string, it's
treated as the head.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>