controller: db: add wifiChannel getter and setter
In the previous commit, a wifiChannel variable was
added to the controller node class.

In this commit, a getter and setter on wifiChannel object
are added in the controller's database module.

In the next commits, these getter and setter will be
applied in the controller's tasks.
The variables that a wifiChannel object replaced will be removed.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>