controller: node: add wifiChannel variable
This is a preparative commit.

Currently, the node class contains multiple channel-related variables.
Replace those variables with wifiChannel variable that groups all
channel-related info.

In the next commits, the usage of this variable will be applied in
the controller database, and in controller tasks.
After that, the current channel-related variables will be removed.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>