bwl: nl80211: use set_create_value when assigning auth and encr params
There are cases where we not only want to assign the authentication
and encryption parameters to vaps, but also to the head.

To make it possible to do that transparently in
assign_auth_encr_parameters(), use the newly-introduced
set_create_vap_value(), to assign parameters no matter if it's for the
head or a vap.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>