Agent: Add a way to configure profile_x_disallow_override in conf file
Add a way to configure the traffic separation static member
"profile_x_disallow_override" using the beerocks_agent.conf file.

Add the configuration to the ".conf" file and to the relevant places
in the code to read it.
Set the configuration on the traffic separation on the Agent main file.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>