Agent: Fix a wrong creation of VLAN interfaces
The traffic separation class creates VLAN interfaces for each interface
that contains the main BSS name.
For instance, for wlan2.2, b0_wlan2.2 b1_wlan2.2 it will create the
interfaces wlan2.2.pvid, b0_wlan2.2.pvid b1_wlan2.2.pvid.

The problem is that on the second running of the code, the new
interfaces will also be considered as interfaces containing the
interface name, thus it will try to add new interfaces with ".pvid"
suffix. I.E. "wlan2.2.pvid.pvid".

Obviously, this is wrong.
This commit removes interfaces that end with ".pvid" so a new VLAN
interface will not be added for them.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>