bwl: add new method to get keys for a station
In the future this method is going to be used by the VBSS feature to
get the key and key sequences for a station when it's about to be
moved from one Agent to another one.

For now no BWL implemented it.

In a follow-up commit, an implementation based on nl80211_client will
be added.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>