controller: fix task id not set when creating vbss_task
Currently, triggering a VBSS move leads to:
Could not trigger VBSS move! VBSS Task not found!

The only place where the VBSS task id is assigned to in the agent DB
is in the vbss_task constructor.
However, in this constructor there is also a check that the task id is
already set. The intention was most likely to check that the task id
was NOT set already, so reverse the condition.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>