agent: ap_manager: send Virtual BSS Response when handling the request
The VBSS specifications mention that the Virtual BSS Response should
be sent when the Virtual BSS Request is handled by the agent, to let
the controller know of what happened.

It was not added when handle_virtual_bss_request() was introduced, so
add it now.

From the VBSS task, send it to the controller.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>