agent: ap_manager: refresh vaps info after a BSS has been added/removed
According to the VBSS specification once the Virtual BSS Request
message has been handled successfully, a topology notification must be
sent to make the controller aware of the changes.

From the topology notification, the controller will send a topology
query, and will see from the response that new interfaces have been
created. From there, it's supposed to use AP_METRICS_QUERY_MESSAGE to
get more information about them.

Unfortunately, prplMesh does it differently and instead uses a
vendor-specific message to give the controller the list of vaps. In
addition, the list of vaps prplMesh includes in the topology response
message is currently not correct (it assumes one fronthaul and
backhaul interface per radio maximum).

Since we can't get rid of the vendor specific flow as part of the
current changes, instead of only sending a topology notification send
ACTION_APMANAGER_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION. This will
refresh the list of vaps and also send a topology notification to the
controller.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>