controller: update_vap() on hostap list updates notification
Currently, ACTION_CONTROL_HOSTAP_VAPS_LIST_UPDATE_NOTIFICATION is the
way to receive the complete information about the vaps on a specific
radio of an agent (because the topology response information is
incomplete due to the way the radio is represented in the agent DB).

However, up to now it did not call update_vap() to try to add/update
vaps and also add them to the datamodel. Because of this, when a new
BSS was part of the notification, it was not added to the datamodel.

To fix it, use update_vap which will not only call add_virtual_node()
like it was done before, but also add the new vap in the DM by calling
dm_set_radio_bss().

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>